Fix: Fixed type conversion issue where layer_block.py called a method from layer.py and provided a float where int was expected.

+++ widgets/layers/layer.py
@@ -22,7 +22,7 @@ class Layer(QWidget):
         self.update()
         self.layer_updated.emit()
     
-    def updateOpacity(self, opacity: int):
+    def updateOpacity(self, opacity: float):
         self.opacity = opacity
         self.layer_updated.emit()
